get_OntologyNode

get_OntologyNode

get_OntologyNode(
  compendium = "vespucci",
  normalization = "tpm",
  originalId = "PO_0009005",
  ontology_Name = "Plant ontology"
)

Arguments

compendium
A string - the selected compendium

normalization
A string ('legacy' as default)

originalId
A string - the ontology term

ontology_Name
A string - the name of the selected ontology

Value

A data.frame

Examples

get_OntologyNode(normalization="tpm",originalId="PO_0009005",ontology_Name="Plant ontology")

#> [[1]]
#> [[1]]$node
#> [[1]]$node$id
#> [1] "T250b2xvZ3lOb2RlVHlwZToyMzg4Nzk2"
#> 
#> [[1]]$node$originalId
#> [1] "PO_0009005"
#> 
#> [[1]]$node$termShortName
#> [1] "root"
#> 
#> [[1]]$node$json
#> [[1]]$node$json[[1]]
#> [[1]]$node$json[[1]]$`@id`
#> [1] "http://purl.obolibrary.org/obo/po.owl#PO_0009005"
#> 
#> [[1]]$node$json[[1]]$`http://www.w3.org/2002/07/owl#id`
#> [[1]]$node$json[[1]]$`http://www.w3.org/2002/07/owl#id`[[1]]
#>       @value 
#> "PO:0009005" 
#> 
#> 
#> [[1]]$node$json[[1]]$`http://www.w3.org/2002/07/owl#hasDbXref`
#> [[1]]$node$json[[1]]$`http://www.w3.org/2002/07/owl#hasDbXref`[[1]]
#>       @value 
#> "PO_GIT:578" 
#> 
#> 
#> [[1]]$node$json[[1]]$`http://www.w3.org/2000/01/rdf-schema#label`
#> [[1]]$node$json[[1]]$`http://www.w3.org/2000/01/rdf-schema#label`[[1]]
#> @value 
#> "root" 
#> 
#> 
#> [[1]]$node$json[[1]]$`http://www.w3.org/2000/01/rdf-schema#comment`
#> [[1]]$node$json[[1]]$`http://www.w3.org/2000/01/rdf-schema#comment`[[1]]
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      @value 
#> "Roots function in the absorption of water and inorganic nutrients, anchoring the plant body to the substrate and supporting it, storage of food and nutrients, and vegetative reproduction. The roots of most vascular plant species enter into symbiosis with soil-borne microorganisms. Roots are usually positively geotropic and found underground, although there are many exceptions such as the aerial roots of orchids. Roots often form secondary thickening from the root lateral meristem (PO:0006308). Commonly thought of as one of the three basic parts of the plant body, along with the shoot axis (PO:0025029) and leaves (PO:0025034)." 
#> 
#> 
#> [[1]]$node$json[[1]]$`http://www.w3.org/2002/07/owl#hasExactSynonym`
#> [[1]]$node$json[[1]]$`http://www.w3.org/2002/07/owl#hasExactSynonym`[[1]]
#>                      @value 
#> "ra&#237z (Spanish, exact)" 
#> 
#> [[1]]$node$json[[1]]$`http://www.w3.org/2002/07/owl#hasExactSynonym`[[2]]
#>          @value 
#> "radix (exact)" 
#> 
#> [[1]]$node$json[[1]]$`http://www.w3.org/2002/07/owl#hasExactSynonym`[[3]]
#>                @value 
#> "9 (Japanese, exact)" 
#> 
#> [[1]]$node$json[[1]]$`http://www.w3.org/2002/07/owl#hasExactSynonym`[[4]]
#>                    @value 
#> "radices (exact, plural)" 
#> 
#> 
#> [[1]]$node$json[[1]]$`http://www.w3.org/2002/07/owl#hasOBONamespace`
#> [[1]]$node$json[[1]]$`http://www.w3.org/2002/07/owl#hasOBONamespace`[[1]]
#>          @value 
#> "plant_anatomy" 
#> 
#> 
#> [[1]]$node$json[[1]]$`http://www.w3.org/2002/07/owl#hasAlternativeId`
#> [[1]]$node$json[[1]]$`http://www.w3.org/2002/07/owl#hasAlternativeId`[[1]]
#>       @value 
#> "PO:0003006" 
#> 
#> 
#> [[1]]$node$json[[1]]$`http://www.w3.org/2002/07/owl#hasNarrowSynonym`
#> [[1]]$node$json[[1]]$`http://www.w3.org/2002/07/owl#hasNarrowSynonym`[[1]]
#>                   @value 
#> "climbing root (narrow)" 
#> 
#> [[1]]$node$json[[1]]$`http://www.w3.org/2002/07/owl#hasNarrowSynonym`[[2]]
#>                 @value 
#> "aerial root (narrow)" 
#> 
#> 
#> 
#> 
#> [[1]]$node$ontology
#>                     id 
#> "T250b2xvZ3lUeXBlOjEy" 
#> 
#> 
#>