Get experimentAccessid and experimentName from sampleName

use get_available_compendia to check all the available compendia

get_experiments(compendium = "vespucci", sampleName = NULL)

Arguments

compendium
A string - the selected compendium

sampleName
A string - if NULL(default) returns all available experiments ids for the selected compendium

Value

A data.frame with experimentAccessId and esperimentName

Examples

get_experiments()

#>     experimentAccessId
#> 1          PRJNA596533
#> 2          PRJNA528770
#> 3          PRJNA491334
#> 4          PRJNA529179
#> 5          PRJNA594635
#> 6          E-MEXP-3803
#> 7          PRJNA507625
#> 8          PRJNA265039
#> 9          PRJNA507550
#> 10         PRJNA477842
#> 11           GSE150581
#> 12           GSE138021
#> 13            GSE66913
#> 14         PRJNA526817
#> 15           PRJDB8743
#> 16         PRJNA477721
#> 17          PRJEB36552
#> 18         PRJNA508364
#> 19             GSE6404
#> 20            GSE31674
#> 21            GSE31675
#> 22            GSE29948
#> 23            GSE31677
#> 24            GSE31660
#> 25            GSE34634
#> 26            GSE12842
#> 27            GSE31594
#> 28            GSE31662
#> 29            GSE31664
#> 30            GSE24561
#> 31         E-MEXP-1282
#> 32            GSE35172
#> 33         E-MTAB-1610
#> 34            GSE42345
#> 35            GSE43043
#> 36            GSE44823
#> 37         PRJNA287490
#> 38            GSE36632
#> 39            GSE40487
#> 40         E-MEXP-3045
#> 41         E-MEXP-1950
#> 42         E-MEXP-1524
#> 43            GSE41633
#> 44            GSE49569
#> 45            GSE20511
#> 46            GSE44213
#> 47         E-MTAB-1634
#> 48         E-MTAB-1514
#> 49         E-MEXP-2541
#> 50            GSE58958
#> 51         E-MTAB-1523
#> 52            GSE62744
#> 53         PRJNA168987
#> 54            GSE52586
#> 55            GSE65969
#> 56            GSE70670
#> 57            GSE53409
#> 58            GSE52167
#> 59            GSE67708
#> 60         PRJNA433817
#> 61            GSE75565
#> 62            GSE66391
#> 63            GSE76000
#> 64            GSE86891
#> 65            GSE80801
#> 66           GSE113225
#> 67           GSE106668
#> 68            GSE89185
#> 69            GSE95532
#> 70            GSE43901
#> 71            GSE62745
#> 72            GSE86551
#> 73            GSE92980
#> 74            GSE97578
#> 75            GSE57669
#> 76         E-MTAB-3062
#> 77         E-MTAB-3899
#> 78           GSE118569
#> 79            GSE77240
#> 80            GSE82317
#> 81            GSE97900
#> 82            GSE97960
#> 83            GSE85812
#> 84            GSE87643
#> 85            GSE59436
#> 86         PRJNA348618
#> 87         PRJNA336478
#> 88         PRJNA369777
#> 89         PRJNA378698
#> 90            GSE67191
#> 91            GSE77895
#> 92         PRJNA476839
#> 93            GSE58653
#> 94           GSE107905
#> 95            GSE33386
#> 96            GSE98873
#> 97            GSE33222
#> 98           GSE122411
#> 99           GSE130386
#> 100           GSE76256
#> 101           GSE99726
#> 102          GSE107573
#> 103          GSE124798
#> 104           GSE89113
#> 105          GSE109065
#> 106          GSE103226
#> 107           GSE54636
#> 108            GSE7677
#> 109          GSE111468
#> 110           GSE52829
#> 111          GSE116740
#> 112          GSE117849
#> 113          GSE126052
#> 114            GSE7394
#> 115            GSE7750
#> 116            GSE7749
#> 117           GSE42315
#> 118            GSE7751
#> 119           GSE42972
#> 120            GSE7679
#> 121           GSE69282
#> 122            GSE7680
#> 123            GSE8445
#> 124           GSE57770
#> 125           GSE83275
#> 126           GSE64153
#> 127            GSE7747
#> 128        PRJNA342391
#> 129        PRJNA380033
#> 130        PRJNA474960
#> 131        PRJNA268857
#> 132        PRJNA528145
#> 133        PRJNA293344
#> 134        PRJNA414966
#> 135        PRJNA307077
#> 136        PRJNA517111
#> 137        PRJNA488534
#> 138        PRJNA448744
#> 139         PRJEB30338
#> 140        PRJNA527184
#> 141        PRJNA539972
#> 142        PRJNA316157
#> 143        PRJNA402079
#> 144        PRJNA398733
#> 145        PRJNA436693
#> 146        PRJNA419810
#> 147        PRJNA516950
#> 148        PRJNA394039
#> 149        PRJNA395634
#> 150        PRJNA306749
#> 151        PRJNA514211
#> 152        PRJNA313243
#> 153        PRJNA275778
#> 154        PRJNA309142
#> 155        PRJNA313234
#> 156           GSE27180
#> 157           GSE32343
#> 158        PRJNA521100
#> 159           GSE78920
#> 160           GSE78798
#> 161           GSE11857
#> 162        PRJNA227009
#> 163           GSE11406
#> 164           GSE17502
#> 165           GSE41423
#> 166           GSE36849
#> 167           GSE89075
#> 168           GSE41206
#> 169           GSE36128
#> 170           GSE36234
#> 171           GSE18596
#> 172           GSE28779
#> 173        E-MTAB-2289
#> 174           GSE42909
#> 175           GSE72421
#> 176           GSE75498
#> 177           GSE92864
#> 178           GSE88765
#> 179          GSE107014
#> 180           GSE85079
#> 181        PRJNA484195
#> 182          PRJDB5807
#> 183        PRJNA307079
#> 184          GSE121146
#> 185           GSE71146
#> 186           GSE42312
#> 187            GSE7758
#> 188           GSE53824
#> 189           GSE62315
#> 190            GSE8389
#> 191           GSE56915
#> 192            GSE7752
#> 193        PRJNA203687
#> 194        PRJNA169607
#> 195         PRJEB11405
#> 196        PRJNA521303
#> 197        PRJNA285072
#> 198        PRJNA258459
#> 199        PRJNA390670
#> 200         PRJEB31325
#>                                                                                                                                                                                               experimentName
#> 1                                                                      Comparative transcriptomic analysis between \030Sanbenti\031 and its bud sport \03011-06-25\031 during grape berry development stages
#> 2                                                                                                                                                                              wine grape leaf transcriptome
#> 3                                                                                             Comparison of Transcriptional Expression Patterns of 'Kyoho' Grapes under a Two-crop-a-year Cultivation System
#> 4                                                    Rpv3-1 Locus Mediated Resistance to Grapevine Downy Mildew is Associated with Specific Host Transcriptional Responses and the Accumulation of Stilbenes
#> 5                                                                                                                                                                                                   Leafroll
#> 6                                                                                                                                                    Auto-grafting in grapevine 3 and 28 days after grafting
#> 7                                                                                                                       Rootstock effects on scion phenotypes in a \030Chambourcin\031 experimental vineyard
#> 8                                                                                                                     Transcriptome analysis of 5 white Italian grapevine varieties during berry development
#> 9                                                                                                             Dissecting the Variations of Transcriptome in Grape Berries Grown under Double Cropping System
#> 10                                               Clone-specific transcript profiling of \030Nebbiolo\031 grape berries unveils environmental responses mediated by sugar and secondary metabolite signalling
#> 11                                                                                                     Transcript profiles in leaves of the salt-tolerant grapevine rootstock 1616C under salt and ER-stress
#> 12                                                                                   RAPID DEHYDRATION OF GRAPE BERRIES DAMPENS THE POST-RIPENING TRANSCRIPTOMIC PROGRAM AND THE METABOLITEPROFILE EVOLUTION
#> 13                                                                                                                                                                                        Grape_Bud_Dormancy
#> 14                                                                                                                                                                            Elsinoe ampelina RNA-Seq reads
#> 15                                                                                                                                                                       Transcriptome analysis in grapevine
#> 16                                                                                                                            Vitis vinifera L. Transcriptome (for circRNA identification under cold stress)
#> 17                                                                                                  Transcriptome investigations supports the role of rootstock-scion interaction in grapevine berry quality
#> 18                                                                                                             SafeGrape: unconventional compounds against powdery mildew to enhance plant defense responses
#> 19                                                                                               V. vinifera 'Cabernet sauvignon' and V. aestivalis 'Norton' innoculated with Erysiphe necator conidiospores
#> 20                                                                                                                                                           Pinot Noir berry transcriptome during ripening.
#> 21                                                                                                                                                     High temperature effect on Cabernet Sauvignon berries
#> 22                                                                                                                                                Expression data from 35S::VvCBF4-overexpressing grapevines
#> 23                                                                                                                                                                   Long-term Salt & Water Stress in Grapes
#> 24                                                                                                                                        Gene expression associated with compatible viral diseases in berry
#> 25                                                                                                                                                                       Sauvignon blanc berry developmental
#> 26                                                                                                                                           Gene expression in grapevine in response to Bois noir infection
#> 27                                                                                                                                                              short term abiotic stress Cabernet Sauvignon
#> 28                                                                                                            Grape skin transcriptome in the berries cultured in vitro treated with exogenous abscisic acid
#> 29                                                                                                            Grape skin transcriptome in the berries grown on the vine treated with exogenous abscisic acid
#> 30                                                                                        Berry skin development in Norton grape: Distinct patterns of transcriptional regulation and flavonoid biosynthesis
#> 31                                                                                                          Transcription profiling of grapevine cultivar Pinot Noir berry development during three seasons.
#> 32                                                                                                                                       Transcripts and metabolites of grape ripening and aroma development
#> 33                                                                                                                      Gene expression during graft union formation in auto- and hetero-grafts of grapevine
#> 34                                                                                                                                                                  Time course of grapevine bud development
#> 35                                                                                           Effect of temperature and irradiation intensity on grapevine berry ripening in Muscat Hamburg fruiting cuttings
#> 36                                                                                                                Perturbation of polyamine catabolism affects normal grape ripening of Trincadeira cultivar
#> 37                                                                     Vitis vinifera cultivar:Dwarf rapid cycling and continous flowering pinot meunier mutant (line ML 1) Transcriptome or Gene expression
#> 38  Co-evolution between Grapevine rupestris stem pitting-associated virus and Vitis vinifera L. induces a decrease in defence responses and physiological performance associated with an increase in photos
#> 39                                                                                 Selective defoliation affects plant growth, fruit transcriptional ripening program and flavonoid metabolism in grapevine.
#> 40                                                                                                                         Sulphur dioxide evokes large scale reprogramming of the grape berry transcriptome
#> 41                                                                                                                                                                               RRM transcriptomic analysis
#> 42                                                                                    Transcription profiling of grapevine cell culture treated with methyly jasmonate, salicylic acids and ethanol controls
#> 43                                                                                                                                                       The plasticity of the grapevine berry transcriptome
#> 44                                                                                                                                                    Plasticity of the ripening process among berry classes
#> 45                                                                                                                        Expression profiling in grapevine berry during development and withering processes
#> 46                                                                                                    Transcriptional responses to water deficit and Xylella fastidiosa (Pierce's disease) in Vitis vinifera
#> 47                                                                                                                                                                   2011_culture in vitro_sugar anthocyanin
#> 48                                                                                                                                                       Response of Vitis vinifera to Eutypa lata infection
#> 49                                                                                                                     Transcription profiling of Vitis vinifera cv. Malbec after exposure to UV-B radiation
#> 50                                                                                                                                                           The genome of the grape powdery mildew pathogen
#> 51                                                                                                                                   Response of grapevine shoot apex expression to grafting with rootstocks
#> 52                                                                                                                      Transcriptome analysis of 5 red Italian grapevine varieties during berry development
#> 53                                                                                                                                                                          Vitis vinifera strain:Pinot noir
#> 54                               Transcript and metabolite analysis of Vitis vinifera cv. Trincadeira berries infected with Botrytis cinerea reveals an activation of a non-sustained plant defense response
#> 55                                  Analysis of the molecular dialogue between gray mold (Botrytis cinerea) and grapevine (Vitis vinifera) reveals a clear shift in defense mechanisms during berry ripening
#> 56                                                                                                     Distinct transcriptome responses to water limitation in isohydric and anisohydric grapevine cultivars
#> 57                                                                    Night heat stress triggered different transcriptomic responses than day stress in green and ripening grapevine (Vitis vinifera) fruits
#> 58                                                                                                                                              Circadian expression profiling in grapevine ripening berries
#> 59                                                                                                                                    Expression analysis of Vitis Vinifera clones differing for compactness
#> 60                                                                                                                                    Transcriptome profiling of grapevine kinase genes under drought stress
#> 61                                                                                                                                         Plasticity of the berry ripening program in a white grape variety
#> 62                                                                                      Water limitation and rootstock genotype interact to alter grape berry metabolism through transcriptome reprogramming
#> 63                                                                          Reducing sampling bias in molecular studies of grapevine fruit ripening: transcriptomic assessment of the density sorting method
#> 64                                                       Functional characterization of a WRKY transcription factor involved in the regulation of vacuolar transport and flavonoid biosynthesis in grapevine
#> 65                                                                                   Comparative genome hybridization between grapevine white-berried somatic variants and ancestral black-berried cultivars
#> 66                                                                                                                 Timing and order of the molecular events marking the onset of berry ripening in grapevine
#> 67                                                                                                  The major origin of seedless grapes is associated with a missense mutation in the MADS-box gene VviAGL11
#> 68                                                                                             Transcriptomic Regulation Of Scion/Rootstock Harmony And Drought Response In The Roots Of Grapevine Rootstock
#> 69                                                                                        Transcriptome-wide identification of novel UV-B- and light modulated flavonol pathway genes controlled by VviMYBF1
#> 70                                                                                                                             Transcriptional changes associated with the fleshless (flb) somatic variation
#> 71                                                                                                                    Transcriptome analysis of 5 white Italian grapevine varieties during berry development
#> 72                                                                       Dissecting the biochemical and transcriptomic effects of locally applied heat stress on developing Cabernet Sauvignon grape berries
#> 73                                                                         Transcriptional Responses to Pre-flowering Leaf Defoliation in Grapevine Berry from Different Growing Sites, Years, and Genotypes
#> 74                                Grapevine field experiments reveal the contribution of genotype, the influence of environment and the effect of their interaction (GXE) on the berry transcriptome [array]
#> 75                                                                                         Transcriptomic response in leaves of two Vitis vinifera L. varieties in field and under greenhouse abiotic stress
#> 76                                                       Transcription profiling by array of Vitis rootstock roots to investigate the effects of N supply on different rootstock/scion genotype combinations
#> 77                                                                                                   Gene expression during wood pathogen infection in the presence of a root biocontrol agent in grapevine.
#> 78                                                                                                                            A galling insect activates plant reproductive programs during gall development
#> 79                                                  Metabolite and transcript profiling of VvERF045 OE transgenic lines elucidate the role of this transcription factor in the grapevine ripening physiology
#> 80                                                                                           Grapevine grafting: scion transcript profiling and defense-related metabolites induced by different rootstocks.
#> 81                        Genome-wide expression profiling of grapevine (Vitis vinifera cv Cabernet Sauvignon) woody stems and asymptomatic leaves during a time course of infection by Neofusicoccum parvum
#> 82                              Grapevine field experiments reveal the contribution of genotype, the influence of environment and the effect of their interaction (G3E) on the berry transcriptome [RNA-seq]
#> 83                                                     Field systems biology of grapevine red blotch reveals the impact of viral infections on the transcriptional and hormonal regulation of berry ripening
#> 84                                                                                        O antigen acts as a shield to delay early plant immune recognition of the pathogenic bacterium, Xylella fastidiosa
#> 85                                                                                                                                     Transcriptome analyses of grape berry skin induced by UV irradiation.
#> 86                                                                                                                                          Berry response to water deficit in the Merlot grapevine cultivar
#> 87                                                                                                                       Molecular analysis of interaction between the grapevine flower and Botrytis cinerea
#> 88                                                                                                                                                                  NEW INSIGHTS INTO ABA SIGNALING NETWORK.
#> 89                                                                                           Genome wide identification of long noncoding RNAs responsive to Lasiodiplodia theobromae infection in grapevine
#> 90                                                                         Comparative transcriptomics of Central Asian Vitis vinifera accessions reveals distinct defense strategies against powdery mildew
#> 91                                                                                       To a better understanding of the impact of vine nitrogen status on volatile thiols from plot to transcriptome level
#> 92                                                                                                                                                                                Grape leaves Transcriptome
#> 93                                                                                                  Host-based markers for the early stage of infection by the grapevine trunk pathogen Neofusicoccum parvum
#> 94                                                                                                       A cluster of miRNAs regulate MYB repressors to promote anthocyanin accumulation in grapes (RNA-Seq)
#> 95                                                                                                                                                 Unique mRNA expressed during grapevine floral development
#> 96                                                                                                           RNASeq analysis of Vitis vinifera (cv. Sauvignon blanc) grape berries exposed to elevated light
#> 97                                                          Heat stress effects on gene expression in the grape berry. Comparison of pre- and post-véraison berries and microarray versus mRNA-seq datasets.
#> 98                                                                                                            Dissecting the Variations of Transcriptome in Grape Berries Grown under Double Cropping System
#> 99                                                                                                                 Transcriptome profiles of three Muscat table grape cultivars at three developmental stage
#> 100                                                                                                                           Grape skin transcriptome to reveal genes related with resveratrol accumulation
#> 101                                                                                     Sunlight effects on anthocyanins accumulation and associated co-expression gene networks in developing grape berries
#> 102                                                                                               Transcriptome analysis of Envelope and VaWRKY12 overexpressing callus lines generated from Vitis amurensis
#> 103                                                                                                                                                Gene expression variation in grape transgenic calli lines
#> 104                                                                          Integration of transcriptomic and proteomic analyses reveals new insights into the response of grape leaves to high temperature
#> 105                                                                             Shoot transcriptomic response is more sensitive to a 14 day water deficit than root transcriptomic response in Vitis riparia
#> 106                                                                           Dissecting the Variations of Ripening Progression and Flavonoid Metabolism in Grape Berries Grown under Double Cropping System
#> 107                                                                                                                                  Effect of solar ultraviolet radiation on grape berry skin transcriptome
#> 108                                                                                                                   Grape berry developmental series from a vineyard in Willunga, South Australia (WIL-04)
#> 109                                            The MYB5-driven MBW complex recruits a WRKY factor to enhance the expression of targets involved in vacuolar hyper-acidification and trafficking in grapevine
#> 110                                                                                                        The Transcriptomic Regulation of Berry Development is More Important at Night than During the Day
#> 111                                                               Whole-transcriptome microarray analysis of grapevine-Botrytis cinerea interaction during latent infection of berries (\xe2\xacSnoble rot⬝)
#> 112                                                                                      Root transcriptional profile of 1103P and SO4 grapevine rootstock plants grown in presence and absence of magnesium
#> 113                                                                                                                    Effect of different irrigation supply on gene expression of two table grape cultivars
#> 114                                                                                                                                           Gene expression analysis in grape berries at veraison (SIB-05)
#> 115                                                                                      Analysis of gene expression changes involved in diurnal and circadian rhythms in grape berries. (DIU_05_circadian).
#> 116                                                                                                                            Analysis of gene expression in grape berries due to diurnal effects (DIU_05).
#> 117                                                                                                                                                        Rootstock and soil effects on grape transcriptome
#> 118                                                                                                                            Analysis of gene expression in grape berries due to diurnal effects (DIU_06).
#> 119                                The sulfated laminarin triggers a stress transcriptome before priming the SA- and ROSdependent defenses during grapevine's induced resistance against Plasmopara viticola
#> 120                                                                                  Grape berry developmental series comparing irrigation treatments in a vineyard in Wingara, Victoria, Australia (WIN-05)
#> 121                                                                                                                                         Analysis of the origin of parthenocarpy in Corinto Bianco grapes
#> 122                                                                                  Grape berry developmental series comparing pruning treatments from a vineyard in Clare Valley, South Australia (CLA-05)
#> 123                                                                                                            A comparison of gene expression between the skin and flesh tissue of grape berries (CLAsf_05)
#> 124                                                                                            Root transcriptome comparison between grapevines Cabernet Sauvignon, 140 Ruggeri and K51-40 (+/- salt stress)
#> 125                                                                                          Effect of high CO2 levels and low temperature storage on table grapes quality harvested at two maturity stages.
#> 126                                                                                      Microarray Analysis for the Screening of Genes Inducible by Light or Low Temperature in Post-veraison Grape Berries
#> 127                                                                                   Timing of gene expression for inflorescence development and position of green latent buds on grapevine canes (BUD-06).
#> 128                                                                                   Transcriptome profiling provides insights into N responses of grapevine grafted plants depending on rootstock genotype
#> 129                                                                                               Multi-omic characterization of the Vitis vinifera/Plasmopara viticola (Grapevine Downy Mildew) pathosystem
#> 130                                                                                                                                                     Bordeaux Cabernet Sauvignon berry skin transcriptome
#> 131                                                                                                                                                          Vitis vinifera Transcriptome or Gene expression
#> 132                                                                                                                                                 Transcriptional analysis of  the treatment of riboflavin
#> 133                                                                                                                                     Vitis vinifera cultivar:Tempranillo Transcriptome or Gene expression
#> 134                                                                                                                       Molecular analysis of interaction between the grapevine Berry and Botrytis cinerea
#> 135                                                                                                           Vitis vinifera cultivar:Chasselas Blanc and Chasselas Cioutat Transcriptome or Gene expression
#> 136                                                                                                                                                                                           Vitis grafting
#> 137                                                                                                                                                                               Grape bud dormancy RNA-seq
#> 138                                                                                                                                                                            Vitis vinifera cv Tempranillo
#> 139                                                                                                  Comparative transcriptome analyses of grapevine inflorescences of three different developmental stages.
#> 140                                                                                                                   Transcriptome analysis of light effect on anthocyanin accumulation in grape berry skin
#> 141                                                    Differences in berry primary and secondary metabolisms identified by transcriptomic and metabolic profiling of two table grape color somatic variants
#> 142                                                                                                                                                        Vitis vinifera cultivar:Merlot Raw sequence reads
#> 143                                                                                                                                                                         Vitis vinifera leaf freeze shock
#> 144                                                                                                                                                   Vitis vinifera cv Thompson Seedless Raw sequence reads
#> 145                                                                                                                                                                 Vitis vinifera Grape Berry Transcriptome
#> 146                                                                                                                                  Vitis vinifera strain:superior, 3003, Muscat Hamburg Raw sequence reads
#> 147                                                                                                                   Transcriptomic comparison of four vitis species in response to water and salt stresses
#> 148                                                                                   Identification of the molecular regulators of softening and early ripening events in grape berries (Vitis vinifera L.)
#> 149                                                                                                                                                          Vitis vinifera Transcriptome or Gene expression
#> 150                                                                                                                                        Vitis vinifera cultivar:Loureira Transcriptome or Gene expression
#> 151                                                                                                                                                          Vitis vinifera Transcriptome or Gene expression
#> 152                                                                                                                                                           Vitis davidii Transcriptome or Gene expression
#> 153                                                                                                                               Vitis vinifera cultivar:thompson seedless Transcriptome or Gene expression
#> 154                                                                                                                                       Vitis vinifera cultivar:Brachetto Transcriptome or Gene expression
#> 155                                                                                                                                                Vitis vinifera cultivar:Tocai friulano Raw sequence reads
#> 156                                                                                                              Expression data from micropropagated Vitis vinifera when transferred to ex vitro conditions
#> 157                                                                                                                                                          Grapevine response to Planococcus ficus feeding
#> 158                                                                                                                                             Transcriptome profiles of three Muscat table grape cultivars
#> 159                        Transcriptomic network analyses of leaf dehydration responses identify highly connected ABA and ethylene signaling hubs in three grapevine species differing in drought tolerance
#> 160                                                                                                                                       Abscisic Acid Transcriptomic Signaling Varies with Grapevine Organ
#> 161                                                                                        Gene expression patterns associated with grapevine resistance to downy mildew mediated by the Rpv1 and Rpv2 genes
#> 162                                                                                                                                                                 Vitis vinifera strain:Cabernet sauvignon
#> 163                                                                                                                                    Expression data in individual grape berries during ripening initation
#> 164                                                                                                                                                             Photoperiod regulation of grape bud dormancy
#> 165                                                                                      Transcriptomic analysis during heat stress and after the following recovery in grapevine (Vitis vinifera L.) leaves
#> 166                                                                                                            Gene expression in leaves of grapevine plants under abiotic stress conditions: heat and water
#> 167                          Comparative Physiological and Transcriptome Profiling of Grapevine Rootstocks Identified Their Physiological Adaptations and Gene Regulation  Networks behind Drought Tolerance
#> 168                                                                                                                                                   Grape flesh and skin ripening transcriptomic profiling
#> 169                                                                                     The grapevine expression atlas reveals a deep transcriptome shift driving the entire plant into a maturation program
#> 170                                                             Increasing the source/sink ratio in Vitis vinifera (cv Sangiovese) induces extensive transcriptome reprogramming and modifies berry ripening
#> 171                                                                                                     RNAs expression profiling in a resistant and a susceptible grapevine cultivar to Plasmopara viticola
#> 172                                                                            Transcript and metabolite analysis in Trincadeira cultivar reveals novel information regarding the dynamics of grape ripening
#> 173                                                                                                                                           Time course of grapevine tendril and inflorescence development
#> 174                                                                                         Transcriptome analysis along berry development of 'Syrah' x 'Pinot Noir' offspring extreme for flavonoid content
#> 175                                                  Five Omic Technologies Are Concordant in Differentiating the Biochemical Characteristics of the Berries of Five Grapevine (Vitis vinifera L.) Cultivars
#> 176                                                                                                         Disclosing the Molecular Basis of the Postharvest Life of Berry in Different Grapevine Genotypes
#> 177                                                                                             Whole Plant Temperature Manipulation Affects Flavonoid Metabolism and the Transcriptome of Grapevine Berries
#> 178 The Grapevine Vegetative Color Locus Encompasses Three Closely Related R2R3-MYB Transcription Factors and Regulates Anthocyanin Synthesis with Different Specificities Compared to the Berry Color Locus
#> 179                                                                                                Analysis of the origin of stenospermocarpy under the control of the major seedlessness locus in grapevine
#> 180                Genome-wide expression profiling of Neofusicoccum parvum virulence factors in response to different growth substrates in vitro and during an interaction time course with grapevine stems
#> 181                                                                                    Transcriptomic analysis of a white-to-red skin color reversion in grapevine (Vitis vinifera L. cv. ´Moscatel Galego´)
#> 182                                                                                                                                   Transcriptome analysis for shoot of Japanese wine grape cultivar Koshu
#> 183                                                                                                                                   Vitis vinifera cultivar:Prieto Picudo Transcriptome or Gene expression
#> 184                                                        Volatile profiles and transcriptomic variation in Cabernet Sauvignon grapes in response to various sunlight exposure strategies in a warm climate
#> 185                                                                 Using the combined analysis of transcripts and metabolites to propose key genes for differential terpene accumulation across two regions
#> 186                                                                                                           Grapevine cell cultures transcriptomic response to methyl jasmonate and cyclodextrin elicitors
#> 187                                                                                                                                                        Hormonal effects on grape berry ripening (HOM_06)
#> 188                                                                                                                                  Powdery Mildew- and Salicylic Acid-Induced Gene Expression in Grapevine
#> 189                                                                                                           Transcriptomic analysis of grape (Vitis vinifera L.) leaves exposed to ultraviolet C radiation
#> 190                                                                                                                     The effects of heat shock treatment on gene expression of grapevine berries (HSP_06)
#> 191                                                            A grapevine anthocyanin acyltransferase, transcriptionally regulated by VvMYBA, can produce most acylated anthocyanins present in grape skins
#> 192                                                                                                                                           Gene expression analysis in grape berries at veraison (SIB_06)
#> 193                                                                                                                                            Vitis vinifera strain:Tannat Transcriptome or Gene expression
#> 194                                                                                                                                           Vitis vinifera strain:Corvina Transcriptome or Gene expression
#> 195                                                                                              Next Generation Sequencing of grapevine wood pathogen infection in the presence of a root biocontrol agent.
#> 196                                                                                                                                         Vitis vinifera root transcriptome analysis under flooding stress
#> 197                                                                                                                Vitis vinifera cultivar:Corinto Bianco and Pedro Ximenez Transcriptome or Gene expression
#> 198                                                                                                                  Vitis labrusca x Vitis vinifera strain:Labruscan grape Transcriptome or Gene expression
#> 199                                                                                                                                           Vitis vinifera transcriptome in response to Xylella fastidiosa
#> 200                           Grapevine comparative early transcriptome profiling reveals that Flavescence dor\xe9e phytoplasma represses plant responses induced by vector feeding in susceptible varieties

get_experiments(sampleName="GSM671721.ch1")

#>   experimentAccessId                                  experimentName
#> 1           GSE31674 Pinot Noir berry transcriptome during ripening.