get annotations for n samples from the selected compendium

get annotations for n samples from the selected compendium

get_sample_annotation(compendium = "vitis_vinifera", n = 10)

Arguments

compendium
a character - the selected compendium

n
an integer: number of sample to retrieve (default 10)

Value

a data.frame

Examples

get_sample_annotation(n=25)

#>       sampleName                 sampleId                 ontology.originalId
#> 1  GSM287866.ch1 U2FtcGxlVHlwZToxMDE0OA==                                2006
#> 2  GSM287866.ch1 U2FtcGxlVHlwZToxMDE0OA==         GENOTYPE.CABERNET_SAUVIGNON
#> 3  GSM287866.ch1 U2FtcGxlVHlwZToxMDE0OA==                  EICHHORN-LORENZ.33
#> 4  GSM287866.ch1 U2FtcGxlVHlwZToxMDE0OA==       PO:0008002_fruit_proximal_end
#> 5  GSM287866.ch1 U2FtcGxlVHlwZToxMDE0OA==          VINEYARD.OSOYOOS_BC_CANADA
#> 6  GSM299462.ch1 U2FtcGxlVHlwZToxMDA0MQ==                       LEAF_POSITION
#> 7  GSM299462.ch1 U2FtcGxlVHlwZToxMDA0MQ==                           PLANT_AGE
#> 8  GSM299462.ch1 U2FtcGxlVHlwZToxMDA0MQ==                          GREENHOUSE
#> 9  GSM299462.ch1 U2FtcGxlVHlwZToxMDA0MQ==    GENOTYPE.MUSCADINIA_ROTUNDIFOLIA
#> 10 GSM299462.ch1 U2FtcGxlVHlwZToxMDA0MQ==                     PO:0025034_leaf
#> 11 GSM322364.ch1 U2FtcGxlVHlwZToxMDE4Mw==                                2006
#> 12 GSM322364.ch1 U2FtcGxlVHlwZToxMDE4Mw==                   GENOTYPE.MACCABEU
#> 13 GSM322364.ch1 U2FtcGxlVHlwZToxMDE4Mw==                       LEAF_POSITION
#> 14 GSM322364.ch1 U2FtcGxlVHlwZToxMDE4Mw==                PO:0020138_leaf_vein
#> 15 GSM322364.ch1 U2FtcGxlVHlwZToxMDE4Mw== VINEYARD.COVIOLO-REGGIOEMILIA_ITALY
#> 16 GSM322364.ch1 U2FtcGxlVHlwZToxMDE4Mw==                     VINEYARD.UNKOWN
#> 17 GSM322369.ch1 U2FtcGxlVHlwZToxMDE4OA==                                2006
#> 18 GSM322369.ch1 U2FtcGxlVHlwZToxMDE4OA==                    GENOTYPE.MANZONI
#> 19 GSM322369.ch1 U2FtcGxlVHlwZToxMDE4OA==                       LEAF_POSITION
#> 20 GSM322369.ch1 U2FtcGxlVHlwZToxMDE4OA==                PO:0020138_leaf_vein
#> 21 GSM322369.ch1 U2FtcGxlVHlwZToxMDE4OA== VINEYARD.COVIOLO-REGGIOEMILIA_ITALY
#> 22 GSM322369.ch1 U2FtcGxlVHlwZToxMDE4OA==                     VINEYARD.UNKOWN
#> 23 GSM404719.ch2 U2FtcGxlVHlwZToxMTA1OA==                                TIME
#> 24 GSM404719.ch2 U2FtcGxlVHlwZToxMTA1OA==                                2001
#> 25 GSM404719.ch2 U2FtcGxlVHlwZToxMTA1OA==                   GENOTYPE.PERLETTE
#>                  name
#> 1  Condition ontology
#> 2  Condition ontology
#> 3  Condition ontology
#> 4  Condition ontology
#> 5  Condition ontology
#> 6  Condition ontology
#> 7  Condition ontology
#> 8  Condition ontology
#> 9  Condition ontology
#> 10 Condition ontology
#> 11 Condition ontology
#> 12 Condition ontology
#> 13 Condition ontology
#> 14 Condition ontology
#> 15 Condition ontology
#> 16 Condition ontology
#> 17 Condition ontology
#> 18 Condition ontology
#> 19 Condition ontology
#> 20 Condition ontology
#> 21 Condition ontology
#> 22 Condition ontology
#> 23 Condition ontology
#> 24 Condition ontology
#> 25 Condition ontology