Get experimentAccessid and experimentName from sampleName

use get_available_compendia to check all the available compendia

get_experiments(compendium = "vespucci", sampleName = NULL)

Arguments

compendium
A string - the selected compendium

sampleName
A string - if NULL(default) returns all available experiments ids for the selected compendium

Value

A data.frame with experimentAccessId and esperimentName

Examples

get_experiments()

#>    experimentAccessId
#> 1         E-MEXP-3803
#> 2             GSE6404
#> 3            GSE31674
#> 4            GSE31675
#> 5            GSE29948
#> 6            GSE31677
#> 7            GSE31660
#> 8            GSE34634
#> 9            GSE12842
#> 10           GSE31594
#> 11           GSE31662
#> 12           GSE31664
#> 13           GSE24561
#> 14        E-MEXP-1282
#> 15           GSE35172
#> 16        E-MTAB-1610
#> 17           GSE42345
#> 18           GSE43043
#> 19           GSE44823
#> 20        PRJNA287490
#> 21           GSE36632
#> 22           GSE40487
#> 23        E-MEXP-3045
#> 24        E-MEXP-1950
#> 25        E-MEXP-1524
#> 26           GSE41633
#> 27           GSE49569
#> 28           GSE20511
#> 29           GSE44213
#> 30        E-MTAB-1634
#> 31        E-MTAB-1514
#> 32        E-MEXP-2541
#> 33           GSE58958
#> 34        E-MTAB-1523
#> 35           GSE62744
#> 36        PRJNA168987
#> 37           GSE52586
#> 38           GSE65969
#> 39           GSE70670
#> 40           GSE53409
#> 41           GSE52167
#> 42           GSE67708
#> 43        PRJNA433817
#> 44           GSE75565
#> 45           GSE66391
#> 46           GSE76000
#> 47           GSE80801
#> 48           GSE89185
#> 49           GSE43901
#> 50           GSE62745
#> 51           GSE86551
#> 52           GSE97578
#> 53           GSE57669
#> 54        E-MTAB-3062
#> 55           GSE97900
#> 56           GSE97960
#> 57           GSE85812
#> 58        PRJNA378698
#> 59        PRJNA474960
#> 60        PRJNA528145
#> 61        PRJNA293344
#> 62        PRJNA307077
#> 63        PRJNA517111
#> 64        PRJNA448744
#> 65         PRJEB30338
#> 66        PRJNA316157
#> 67        PRJNA398733
#> 68        PRJNA394039
#> 69        PRJNA306749
#> 70        PRJNA514211
#> 71        PRJNA309142
#> 72           GSE27180
#> 73           GSE32343
#> 74           GSE11857
#> 75        PRJNA227009
#> 76           GSE11406
#> 77           GSE17502
#> 78           GSE41423
#> 79           GSE36849
#> 80           GSE89075
#> 81           GSE41206
#> 82           GSE36128
#> 83           GSE36234
#> 84           GSE18596
#> 85           GSE28779
#> 86        E-MTAB-2289
#> 87         E-MTAB-481
#> 88           GSE72421
#> 89           GSE75498
#> 90           GSE85079
#> 91        PRJNA484195
#> 92          PRJDB5807
#> 93        PRJNA307079
#> 94           GSE71146
#> 95        PRJNA203687
#> 96         PRJEB11405
#> 97        PRJNA521303
#> 98        PRJNA285072
#>                                                                                                                                                                                              experimentName
#> 1                                                                                                                                                   Auto-grafting in grapevine 3 and 28 days after grafting
#> 2                                                                                               V. vinifera 'Cabernet sauvignon' and V. aestivalis 'Norton' innoculated with Erysiphe necator conidiospores
#> 3                                                                                                                                                           Pinot Noir berry transcriptome during ripening.
#> 4                                                                                                                                                     High temperature effect on Cabernet Sauvignon berries
#> 5                                                                                                                                                Expression data from 35S::VvCBF4-overexpressing grapevines
#> 6                                                                                                                                                                   Long-term Salt & Water Stress in Grapes
#> 7                                                                                                                                        Gene expression associated with compatible viral diseases in berry
#> 8                                                                                                                                                                       Sauvignon blanc berry developmental
#> 9                                                                                                                                           Gene expression in grapevine in response to Bois noir infection
#> 10                                                                                                                                                             short term abiotic stress Cabernet Sauvignon
#> 11                                                                                                           Grape skin transcriptome in the berries cultured in vitro treated with exogenous abscisic acid
#> 12                                                                                                           Grape skin transcriptome in the berries grown on the vine treated with exogenous abscisic acid
#> 13                                                                                       Berry skin development in Norton grape: Distinct patterns of transcriptional regulation and flavonoid biosynthesis
#> 14                                                                                                         Transcription profiling of grapevine cultivar Pinot Noir berry development during three seasons.
#> 15                                                                                                                                      Transcripts and metabolites of grape ripening and aroma development
#> 16                                                                                                                     Gene expression during graft union formation in auto- and hetero-grafts of grapevine
#> 17                                                                                                                                                                 Time course of grapevine bud development
#> 18                                                                                          Effect of temperature and irradiation intensity on grapevine berry ripening in Muscat Hamburg fruiting cuttings
#> 19                                                                                                               Perturbation of polyamine catabolism affects normal grape ripening of Trincadeira cultivar
#> 20                                                                    Vitis vinifera cultivar:Dwarf rapid cycling and continous flowering pinot meunier mutant (line ML 1) Transcriptome or Gene expression
#> 21 Co-evolution between Grapevine rupestris stem pitting-associated virus and Vitis vinifera L. induces a decrease in defence responses and physiological performance associated with an increase in photos
#> 22                                                                                Selective defoliation affects plant growth, fruit transcriptional ripening program and flavonoid metabolism in grapevine.
#> 23                                                                                                                        Sulphur dioxide evokes large scale reprogramming of the grape berry transcriptome
#> 24                                                                                                                                                                              RRM transcriptomic analysis
#> 25                                                                                   Transcription profiling of grapevine cell culture treated with methyly jasmonate, salicylic acids and ethanol controls
#> 26                                                                                                                                                      The plasticity of the grapevine berry transcriptome
#> 27                                                                                                                                                   Plasticity of the ripening process among berry classes
#> 28                                                                                                                       Expression profiling in grapevine berry during development and withering processes
#> 29                                                                                                   Transcriptional responses to water deficit and Xylella fastidiosa (Pierce's disease) in Vitis vinifera
#> 30                                                                                                                                                                  2011_culture in vitro_sugar anthocyanin
#> 31                                                                                                                                                      Response of Vitis vinifera to Eutypa lata infection
#> 32                                                                                                                    Transcription profiling of Vitis vinifera cv. Malbec after exposure to UV-B radiation
#> 33                                                                                                                                                          The genome of the grape powdery mildew pathogen
#> 34                                                                                                                                  Response of grapevine shoot apex expression to grafting with rootstocks
#> 35                                                                                                                     Transcriptome analysis of 5 red Italian grapevine varieties during berry development
#> 36                                                                                                                                                                         Vitis vinifera strain:Pinot noir
#> 37                              Transcript and metabolite analysis of Vitis vinifera cv. Trincadeira berries infected with Botrytis cinerea reveals an activation of a non-sustained plant defense response
#> 38                                 Analysis of the molecular dialogue between gray mold (Botrytis cinerea) and grapevine (Vitis vinifera) reveals a clear shift in defense mechanisms during berry ripening
#> 39                                                                                                    Distinct transcriptome responses to water limitation in isohydric and anisohydric grapevine cultivars
#> 40                                                                   Night heat stress triggered different transcriptomic responses than day stress in green and ripening grapevine (Vitis vinifera) fruits
#> 41                                                                                                                                             Circadian expression profiling in grapevine ripening berries
#> 42                                                                                                                                   Expression analysis of Vitis Vinifera clones differing for compactness
#> 43                                                                                                                                   Transcriptome profiling of grapevine kinase genes under drought stress
#> 44                                                                                                                                        Plasticity of the berry ripening program in a white grape variety
#> 45                                                                                     Water limitation and rootstock genotype interact to alter grape berry metabolism through transcriptome reprogramming
#> 46                                                                         Reducing sampling bias in molecular studies of grapevine fruit ripening: transcriptomic assessment of the density sorting method
#> 47                                                                                  Comparative genome hybridization between grapevine white-berried somatic variants and ancestral black-berried cultivars
#> 48                                                                                            Transcriptomic Regulation Of Scion/Rootstock Harmony And Drought Response In The Roots Of Grapevine Rootstock
#> 49                                                                                                                            Transcriptional changes associated with the fleshless (flb) somatic variation
#> 50                                                                                                                   Transcriptome analysis of 5 white Italian grapevine varieties during berry development
#> 51                                                                      Dissecting the biochemical and transcriptomic effects of locally applied heat stress on developing Cabernet Sauvignon grape berries
#> 52                               Grapevine field experiments reveal the contribution of genotype, the influence of environment and the effect of their interaction (GXE) on the berry transcriptome [array]
#> 53                                                                                        Transcriptomic response in leaves of two Vitis vinifera L. varieties in field and under greenhouse abiotic stress
#> 54                                                      Transcription profiling by array of Vitis rootstock roots to investigate the effects of N supply on different rootstock/scion genotype combinations
#> 55                       Genome-wide expression profiling of grapevine (Vitis vinifera cv Cabernet Sauvignon) woody stems and asymptomatic leaves during a time course of infection by Neofusicoccum parvum
#> 56                             Grapevine field experiments reveal the contribution of genotype, the influence of environment and the effect of their interaction (G3E) on the berry transcriptome [RNA-seq]
#> 57                                                    Field systems biology of grapevine red blotch reveals the impact of viral infections on the transcriptional and hormonal regulation of berry ripening
#> 58                                                                                          Genome wide identification of long noncoding RNAs responsive to Lasiodiplodia theobromae infection in grapevine
#> 59                                                                                                                                                     Bordeaux Cabernet Sauvignon berry skin transcriptome
#> 60                                                                                                                                                 Transcriptional analysis of  the treatment of riboflavin
#> 61                                                                                                                                     Vitis vinifera cultivar:Tempranillo Transcriptome or Gene expression
#> 62                                                                                                           Vitis vinifera cultivar:Chasselas Blanc and Chasselas Cioutat Transcriptome or Gene expression
#> 63                                                                                                                                                                                           Vitis grafting
#> 64                                                                                                                                                                            Vitis vinifera cv Tempranillo
#> 65                                                                                                  Comparative transcriptome analyses of grapevine inflorescences of three different developmental stages.
#> 66                                                                                                                                                        Vitis vinifera cultivar:Merlot Raw sequence reads
#> 67                                                                                                                                                   Vitis vinifera cv Thompson Seedless Raw sequence reads
#> 68                                                                                   Identification of the molecular regulators of softening and early ripening events in grape berries (Vitis vinifera L.)
#> 69                                                                                                                                        Vitis vinifera cultivar:Loureira Transcriptome or Gene expression
#> 70                                                                                                                                                          Vitis vinifera Transcriptome or Gene expression
#> 71                                                                                                                                       Vitis vinifera cultivar:Brachetto Transcriptome or Gene expression
#> 72                                                                                                              Expression data from micropropagated Vitis vinifera when transferred to ex vitro conditions
#> 73                                                                                                                                                          Grapevine response to Planococcus ficus feeding
#> 74                                                                                        Gene expression patterns associated with grapevine resistance to downy mildew mediated by the Rpv1 and Rpv2 genes
#> 75                                                                                                                                                                 Vitis vinifera strain:Cabernet sauvignon
#> 76                                                                                                                                    Expression data in individual grape berries during ripening initation
#> 77                                                                                                                                                             Photoperiod regulation of grape bud dormancy
#> 78                                                                                      Transcriptomic analysis during heat stress and after the following recovery in grapevine (Vitis vinifera L.) leaves
#> 79                                                                                                            Gene expression in leaves of grapevine plants under abiotic stress conditions: heat and water
#> 80                          Comparative Physiological and Transcriptome Profiling of Grapevine Rootstocks Identified Their Physiological Adaptations and Gene Regulation  Networks behind Drought Tolerance
#> 81                                                                                                                                                   Grape flesh and skin ripening transcriptomic profiling
#> 82                                                                                     The grapevine expression atlas reveals a deep transcriptome shift driving the entire plant into a maturation program
#> 83                                                             Increasing the source/sink ratio in Vitis vinifera (cv Sangiovese) induces extensive transcriptome reprogramming and modifies berry ripening
#> 84                                                                                                     RNAs expression profiling in a resistant and a susceptible grapevine cultivar to Plasmopara viticola
#> 85                                                                            Transcript and metabolite analysis in Trincadeira cultivar reveals novel information regarding the dynamics of grape ripening
#> 86                                                                                                                                           Time course of grapevine tendril and inflorescence development
#> 87                                                                                                                                      Transcriptional analysis of late ripening stages of grapevine berry
#> 88                                                  Five Omic Technologies Are Concordant in Differentiating the Biochemical Characteristics of the Berries of Five Grapevine (Vitis vinifera L.) Cultivars
#> 89                                                                                                         Disclosing the Molecular Basis of the Postharvest Life of Berry in Different Grapevine Genotypes
#> 90                Genome-wide expression profiling of Neofusicoccum parvum virulence factors in response to different growth substrates in vitro and during an interaction time course with grapevine stems
#> 91                                                                                    Transcriptomic analysis of a white-to-red skin color reversion in grapevine (Vitis vinifera L. cv. ´Moscatel Galego´)
#> 92                                                                                                                                   Transcriptome analysis for shoot of Japanese wine grape cultivar Koshu
#> 93                                                                                                                                   Vitis vinifera cultivar:Prieto Picudo Transcriptome or Gene expression
#> 94                                                                 Using the combined analysis of transcripts and metabolites to propose key genes for differential terpene accumulation across two regions
#> 95                                                                                                                                            Vitis vinifera strain:Tannat Transcriptome or Gene expression
#> 96                                                                                              Next Generation Sequencing of grapevine wood pathogen infection in the presence of a root biocontrol agent.
#> 97                                                                                                                                         Vitis vinifera root transcriptome analysis under flooding stress
#> 98                                                                                                                Vitis vinifera cultivar:Corinto Bianco and Pedro Ximenez Transcriptome or Gene expression

get_experiments(sampleName="GSM671721.ch1")

#>   experimentAccessId                                  experimentName
#> 1           GSE31674 Pinot Noir berry transcriptome during ripening.