Get experimentAccessid and experimentName from sampleName

use get_available_compendia to check all the available compendia

get_experiments(compendium = "vespucci", sampleName = NULL)

Arguments

compendium
A string - the selected compendium

sampleName
A string - if NULL(default) returns all available experiments ids for the selected compendium

Value

A data.frame with experimentAccessId and esperimentName

Examples

get_experiments()

#>     experimentAccessId
#> 1            GSE113223
#> 2             GSE66913
#> 3          PRJNA528770
#> 4          PRJNA321480
#> 5          PRJNA307074
#> 6          PRJNA260535
#> 7          PRJNA596533
#> 8            GSE113224
#> 9          PRJNA491334
#> 10           GSE116581
#> 11         PRJNA529179
#> 12           GSE116741
#> 13          PRJEB24540
#> 14            GSE13801
#> 15         PRJNA594635
#> 16         E-MEXP-3803
#> 17         PRJNA589718
#> 18         PRJNA513863
#> 19         PRJNA507625
#> 20         PRJNA398705
#> 21         PRJNA398136
#> 22         PRJNA265039
#> 23            GSE14276
#> 24         PRJNA507550
#> 25            GSE28851
#> 26            GSE10906
#> 27             GSE3052
#> 28            GSE54347
#> 29            GSE52540
#> 30            GSE87296
#> 31            GSE55473
#> 32            GSE55474
#> 33            GSE57588
#> 34         PRJNA269983
#> 35             GSE7681
#> 36            GSE83494
#> 37             GSE9903
#> 38         PRJNA207277
#> 39          PRJEB20991
#> 40         PRJNA397021
#> 41         PRJNA158013
#> 42          PRJEB30496
#> 43             GSE6404
#> 44            GSE31674
#> 45            GSE31675
#> 46            GSE29948
#> 47            GSE58061
#> 48            GSE31677
#> 49            GSE31660
#> 50            GSE34634
#> 51            GSE12842
#> 52            GSE31594
#> 53            GSE31662
#> 54            GSE31664
#> 55            GSE24561
#> 56         E-MEXP-1282
#> 57            GSE35172
#> 58         E-MTAB-1610
#> 59            GSE42345
#> 60            GSE43043
#> 61            GSE44823
#> 62         PRJNA287490
#> 63            GSE36632
#> 64            GSE40487
#> 65         E-MEXP-3045
#> 66         E-MEXP-1950
#> 67         E-MEXP-1524
#> 68            GSE41633
#> 69            GSE49569
#> 70            GSE20511
#> 71            GSE44213
#> 72            GSE16173
#> 73         E-MTAB-1634
#> 74         E-MTAB-1514
#> 75         E-MEXP-2541
#> 76            GSE63512
#> 77            GSE37743
#> 78            GSE58958
#> 79         E-MTAB-1523
#> 80         E-MEXP-2102
#> 81         E-MEXP-2337
#> 82            GSE62744
#> 83         PRJNA168987
#> 84            GSE56844
#> 85            GSE52586
#> 86            GSE65969
#> 87            GSE70670
#> 88            GSE53409
#> 89            GSE52167
#> 90            GSE67708
#> 91         PRJNA433817
#> 92            GSE75565
#> 93            GSE66391
#> 94            GSE76000
#> 95            GSE86891
#> 96            GSE80801
#> 97           GSE113225
#> 98           GSE106668
#> 99            GSE89185
#> 100           GSE43901
#> 101           GSE62745
#> 102           GSE76834
#> 103           GSE86551
#> 104           GSE92980
#> 105        PRJNA415169
#> 106           GSE97578
#> 107           GSE57669
#> 108           GSE86163
#> 109        E-MTAB-3062
#> 110           GSE77218
#> 111          GSE118569
#> 112           GSE97900
#> 113           GSE97960
#> 114           GSE85812
#> 115           GSE87643
#> 116        PRJNA348618
#> 117        PRJNA378698
#> 118        PRJNA378596
#> 119           GSE77895
#> 120        PRJNA476839
#> 121           GSE58653
#> 122           GSE98873
#> 123          GSE122411
#> 124           GSE76256
#> 125          GSE109065
#> 126          GSE103226
#> 127          GSE116740
#> 128          GSE126052
#> 129            GSE7749
#> 130           GSE42315
#> 131            GSE7679
#> 132           GSE57770
#> 133           GSE83275
#> 134        PRJNA260645
#> 135        PRJNA187271
#> 136        PRJNA380033
#> 137        PRJNA474960
#> 138        PRJNA268857
#> 139        PRJNA283438
#> 140        PRJNA528145
#> 141        PRJNA255837
#> 142        PRJNA338255
#> 143        PRJNA293344
#> 144        PRJNA269987
#> 145        PRJNA414966
#> 146        PRJNA307077
#> 147        PRJNA517111
#> 148        PRJNA488534
#> 149        PRJNA448744
#> 150         PRJEB30338
#> 151        PRJNA472397
#> 152        PRJNA464034
#> 153        PRJNA527184
#> 154        PRJNA525460
#> 155        PRJNA539972
#> 156        PRJNA316157
#> 157        PRJNA448981
#> 158        PRJNA402079
#> 159        PRJNA398733
#> 160        PRJNA515329
#> 161        PRJNA436693
#> 162        PRJNA419810
#> 163        PRJNA434655
#> 164        PRJNA516950
#> 165        PRJNA394039
#> 166        PRJNA395634
#> 167        PRJNA306749
#> 168        PRJNA514211
#> 169        PRJNA471579
#> 170        PRJNA313243
#> 171        PRJNA275778
#> 172        PRJNA309142
#> 173        PRJNA286165
#> 174           GSE27180
#> 175           GSE32343
#> 176        PRJNA521100
#> 177        PRJNA312428
#> 178           GSE11857
#> 179        PRJNA227009
#> 180           GSE11406
#> 181           GSE17502
#> 182           GSE41423
#> 183           GSE36849
#> 184           GSE89075
#> 185           GSE41206
#> 186           GSE36128
#> 187           GSE36234
#> 188           GSE18596
#> 189           GSE28779
#> 190        PRJNA347533
#> 191        E-MTAB-2289
#> 192         E-MTAB-481
#> 193           GSE42909
#> 194           GSE72421
#> 195           GSE75498
#> 196           GSE85079
#> 197           GSE98923
#> 198        PRJNA484195
#> 199          PRJDB5807
#> 200        PRJNA307079
#> 201          GSE121146
#> 202           GSE71146
#> 203           GSE42312
#> 204        PRJNA350390
#> 205            GSE8389
#> 206        PRJNA203687
#> 207        PRJNA412790
#> 208        PRJNA169607
#> 209         PRJEB11405
#> 210        PRJNA521303
#> 211        PRJNA285072
#> 212        PRJNA258459
#> 213        PRJNA390670
#> 214         PRJEB31325
#> 215           GSE28439
#>                                                                                                                                                                                               experimentName
#> 1                                                                                                      Timing and order of the molecular events marking the onset of berry ripening in grapevine [WRKY19 OE]
#> 2                                                                                                                                                                                         Grape_Bud_Dormancy
#> 3                                                                                                                                                                              wine grape leaf transcriptome
#> 4                                                                                                                                                                                  Pinot clones resequencing
#> 5                                                                                                           Vitis vinifera cultivar:Red Globe x Crimson Seedless F1 progeny Transcriptome or Gene expression
#> 6                                                 Vitis vinifera cultivar:Cabernet Franc, Cabernet Sauvignon, Chardonnay, Merlot, Pinot Noir, Sauvignon Blanc, and Semillon Transcriptome or Gene expression
#> 7                                                                      Comparative transcriptomic analysis between \030Sanbenti\031 and its bud sport \03011-06-25\031 during grape berry development stages
#> 8                                                                                                        Timing and order of the molecular events marking the onset of berry ripening in grapevine [BHLH OE]
#> 9                                                                                             Comparison of Transcriptional Expression Patterns of 'Kyoho' Grapes under a Two-crop-a-year Cultivation System
#> 10                                                                                                                                           fruitENCODE: An encyclopedia of DNA elements for fruit ripening
#> 11                                                   Rpv3-1 Locus Mediated Resistance to Grapevine Downy Mildew is Associated with Specific Host Transcriptional Responses and the Accumulation of Stilbenes
#> 12                                                                                                 Whole-transcriptome analysis of grapevine-Botrytis cinerea interaction during latent infection of berries
#> 13                                                                      New perspectives in sustainable viticulture: unique resistant traits against downy mildew in the Vitis vinifera domestication center
#> 14                                                                                                                                                    Grape miRNAs expression profiling in different organs.
#> 15                                                                                                                                                                                                  Leafroll
#> 16                                                                                                                                                   Auto-grafting in grapevine 3 and 28 days after grafting
#> 17                                                                                                                                                           Vitis vinifera Transcriptome or Gene expression
#> 18                                                                                                                    Gene regulation in wild and cultivated Vitis species during deacclimation and budbreak
#> 19                                                                                                                      Rootstock effects on scion phenotypes in a \030Chambourcin\031 experimental vineyard
#> 20                                                                                             RNAseq-based genome annotation and identification of long-noncoding RNAs in the grapevine cultivar 'Riesling'
#> 21                                                                                                                                                                         Vitis vinifera Raw sequence reads
#> 22                                                                                                                    Transcriptome analysis of 5 white Italian grapevine varieties during berry development
#> 23                                                                                                        Generation of transcriptome catalogs for global analysis of gene expression in non-model organisms
#> 24                                                                                                            Dissecting the Variations of Transcriptome in Grape Berries Grown under Double Cropping System
#> 25                                       Resistance to Plasmopara viticola in a grapevine segregating population is associated with stilbenoid accumulation and with specific host transcriptional responses
#> 26                                                                                          'Bois noir' phytoplasma induces significant reprogramming of the leaf transcriptome in the field grown grapevine
#> 27                                                                                                                                     Photoperiod induction of endodormancy in Vitis riparia axillary buds.
#> 28                                                                                                                      Identification of grape genes regulated in response to resistance inducer treatments
#> 29                                            A microarray analysis highlights the role of tetrapyrrole pathways in grapevine responses to stolbur phytoplasma, phloem virus infections and recovered status
#> 30                                                                                                                                             cold stress -Impact of cold stress on grapevine inflorescence
#> 31                                                                                    High-throughput SuperSAGE for transcript profiling in berry tissues of Vitis vinifera L. under different light regimes
#> 32                                                                         Functional characterization of a new grapevine MYB transcription factor and regulation of proanthocyanidin biosynthesis in grapes
#> 33                                                                           Genome-wide transcriptional profiling of Botrytis cinerea genes targeting plant cell walls during infections of different hosts
#> 34                                                                                                                               Vitis pseudoreticulata cultivar:Baihe-13-1 Transcriptome or Gene expression
#> 35                                                                                                                              Grape berry expression profiling: developmental series and treatment effects
#> 36                                                                                            Metabolite and transcript profile of grapevine clusters at 6th,7th,8th WAA and the effect of 1-MCP on ripening
#> 37                                                                                                                                      Chilling requirement fulfillment in Vitis riparia overwintering buds
#> 38                                                                                                                               Find new viroids through deep sequencing of Malus pumila and Vitis vinifera
#> 39                                                                                                                                                             Epigenetic diversity in Barossa Shiraz grapes
#> 40                                                                                                                                                                     Vitis rotundifolia Raw sequence reads
#> 41                                                                                                                                                                                        Vitis rotundifolia
#> 42                                                                                                          Seasonal microbial dynamics on grapevine leaves under biocontrol and copper fungicide treatments
#> 43                                                                                               V. vinifera 'Cabernet sauvignon' and V. aestivalis 'Norton' innoculated with Erysiphe necator conidiospores
#> 44                                                                                                                                                           Pinot Noir berry transcriptome during ripening.
#> 45                                                                                                                                                     High temperature effect on Cabernet Sauvignon berries
#> 46                                                                                                                                                Expression data from 35S::VvCBF4-overexpressing grapevines
#> 47                       Transcriptome analysis during seed development provides insights into co-regulated and altered gene expression between a seeded wine grape variety and it seedless somatic variant.
#> 48                                                                                                                                                                   Long-term Salt & Water Stress in Grapes
#> 49                                                                                                                                        Gene expression associated with compatible viral diseases in berry
#> 50                                                                                                                                                                       Sauvignon blanc berry developmental
#> 51                                                                                                                                           Gene expression in grapevine in response to Bois noir infection
#> 52                                                                                                                                                              short term abiotic stress Cabernet Sauvignon
#> 53                                                                                                            Grape skin transcriptome in the berries cultured in vitro treated with exogenous abscisic acid
#> 54                                                                                                            Grape skin transcriptome in the berries grown on the vine treated with exogenous abscisic acid
#> 55                                                                                        Berry skin development in Norton grape: Distinct patterns of transcriptional regulation and flavonoid biosynthesis
#> 56                                                                                                          Transcription profiling of grapevine cultivar Pinot Noir berry development during three seasons.
#> 57                                                                                                                                       Transcripts and metabolites of grape ripening and aroma development
#> 58                                                                                                                      Gene expression during graft union formation in auto- and hetero-grafts of grapevine
#> 59                                                                                                                                                                  Time course of grapevine bud development
#> 60                                                                                           Effect of temperature and irradiation intensity on grapevine berry ripening in Muscat Hamburg fruiting cuttings
#> 61                                                                                                                Perturbation of polyamine catabolism affects normal grape ripening of Trincadeira cultivar
#> 62                                                                     Vitis vinifera cultivar:Dwarf rapid cycling and continous flowering pinot meunier mutant (line ML 1) Transcriptome or Gene expression
#> 63  Co-evolution between Grapevine rupestris stem pitting-associated virus and Vitis vinifera L. induces a decrease in defence responses and physiological performance associated with an increase in photos
#> 64                                                                                 Selective defoliation affects plant growth, fruit transcriptional ripening program and flavonoid metabolism in grapevine.
#> 65                                                                                                                         Sulphur dioxide evokes large scale reprogramming of the grape berry transcriptome
#> 66                                                                                                                                                                               RRM transcriptomic analysis
#> 67                                                                                    Transcription profiling of grapevine cell culture treated with methyly jasmonate, salicylic acids and ethanol controls
#> 68                                                                                                                                                       The plasticity of the grapevine berry transcriptome
#> 69                                                                                                                                                    Plasticity of the ripening process among berry classes
#> 70                                                                                                                        Expression profiling in grapevine berry during development and withering processes
#> 71                                                                                                    Transcriptional responses to water deficit and Xylella fastidiosa (Pierce's disease) in Vitis vinifera
#> 72                                                                                                              Gene expression profiling of grape-bud response to two alternative dormancy-release stimuli.
#> 73                                                                                                                                                                   2011_culture in vitro_sugar anthocyanin
#> 74                                                                                                                                                       Response of Vitis vinifera to Eutypa lata infection
#> 75                                                                                                                     Transcription profiling of Vitis vinifera cv. Malbec after exposure to UV-B radiation
#> 76                                                                                                       Transcriptome Comparison of Cabernet Sauvignon Grape Berries from Two Regions with Distinct Climate
#> 77                                                  Genome-wide analysis of the grapevine stilbene synthase multigenic family: genomic organization and expression profiles upon biotic and abiotic stresses
#> 78                                                                                                                                                           The genome of the grape powdery mildew pathogen
#> 79                                                                                                                                   Response of grapevine shoot apex expression to grafting with rootstocks
#> 80                                                                              Transcription profiling of Cabernet Sauvignon grape vine leaves from plantlets infected with the fungal pathogen Eutypa lata
#> 81                     Transcription profiling of Vitis vinifera cv. Cabernet-Sauvignon leaves obtained from plants grown in different enviroments and infected with vascular ascomycete fungus Eutypa lata.
#> 82                                                                                                                      Transcriptome analysis of 5 red Italian grapevine varieties during berry development
#> 83                                                                                                                                                                          Vitis vinifera strain:Pinot noir
#> 84                                                                          Next Generation Sequencing Facilitates Quantitative Analysis of Vitis flower Wild Type and the domesticated Vitis Transcriptomes
#> 85                               Transcript and metabolite analysis of Vitis vinifera cv. Trincadeira berries infected with Botrytis cinerea reveals an activation of a non-sustained plant defense response
#> 86                                  Analysis of the molecular dialogue between gray mold (Botrytis cinerea) and grapevine (Vitis vinifera) reveals a clear shift in defense mechanisms during berry ripening
#> 87                                                                                                     Distinct transcriptome responses to water limitation in isohydric and anisohydric grapevine cultivars
#> 88                                                                    Night heat stress triggered different transcriptomic responses than day stress in green and ripening grapevine (Vitis vinifera) fruits
#> 89                                                                                                                                              Circadian expression profiling in grapevine ripening berries
#> 90                                                                                                                                    Expression analysis of Vitis Vinifera clones differing for compactness
#> 91                                                                                                                                    Transcriptome profiling of grapevine kinase genes under drought stress
#> 92                                                                                                                                         Plasticity of the berry ripening program in a white grape variety
#> 93                                                                                      Water limitation and rootstock genotype interact to alter grape berry metabolism through transcriptome reprogramming
#> 94                                                                          Reducing sampling bias in molecular studies of grapevine fruit ripening: transcriptomic assessment of the density sorting method
#> 95                                                       Functional characterization of a WRKY transcription factor involved in the regulation of vacuolar transport and flavonoid biosynthesis in grapevine
#> 96                                                                                   Comparative genome hybridization between grapevine white-berried somatic variants and ancestral black-berried cultivars
#> 97                                                                                                                 Timing and order of the molecular events marking the onset of berry ripening in grapevine
#> 98                                                                                                  The major origin of seedless grapes is associated with a missense mutation in the MADS-box gene VviAGL11
#> 99                                                                                             Transcriptomic Regulation Of Scion/Rootstock Harmony And Drought Response In The Roots Of Grapevine Rootstock
#> 100                                                                                                                            Transcriptional changes associated with the fleshless (flb) somatic variation
#> 101                                                                                                                   Transcriptome analysis of 5 white Italian grapevine varieties during berry development
#> 102                                                                      Drawing links from transcriptome to metabolites: the evolution of aroma in the ripening berry of Moscato bianco (Vitis vinifera L.)
#> 103                                                                      Dissecting the biochemical and transcriptomic effects of locally applied heat stress on developing Cabernet Sauvignon grape berries
#> 104                                                                        Transcriptional Responses to Pre-flowering Leaf Defoliation in Grapevine Berry from Different Growing Sites, Years, and Genotypes
#> 105                                                                                                                                                                                EcoGVCC grapevine viruses
#> 106                               Grapevine field experiments reveal the contribution of genotype, the influence of environment and the effect of their interaction (GXE) on the berry transcriptome [array]
#> 107                                                                                        Transcriptomic response in leaves of two Vitis vinifera L. varieties in field and under greenhouse abiotic stress
#> 108                                                                                                                                           Hydrogen cyanamide breaks grapevine bud dormancy in the summer
#> 109                                                      Transcription profiling by array of Vitis rootstock roots to investigate the effects of N supply on different rootstock/scion genotype combinations
#> 110                                                                                                                        Transcriptomic study of biological networks involving grapevine berry development
#> 111                                                                                                                           A galling insect activates plant reproductive programs during gall development
#> 112                       Genome-wide expression profiling of grapevine (Vitis vinifera cv Cabernet Sauvignon) woody stems and asymptomatic leaves during a time course of infection by Neofusicoccum parvum
#> 113                             Grapevine field experiments reveal the contribution of genotype, the influence of environment and the effect of their interaction (G3E) on the berry transcriptome [RNA-seq]
#> 114                                                    Field systems biology of grapevine red blotch reveals the impact of viral infections on the transcriptional and hormonal regulation of berry ripening
#> 115                                                                                       O antigen acts as a shield to delay early plant immune recognition of the pathogenic bacterium, Xylella fastidiosa
#> 116                                                                                                                                         Berry response to water deficit in the Merlot grapevine cultivar
#> 117                                                                                          Genome wide identification of long noncoding RNAs responsive to Lasiodiplodia theobromae infection in grapevine
#> 118                                                                                                                                 RNA-Seq of young and mid-ripening berries of Gewurztraminer and Riesling
#> 119                                                                                      To a better understanding of the impact of vine nitrogen status on volatile thiols from plot to transcriptome level
#> 120                                                                                                                                                                               Grape leaves Transcriptome
#> 121                                                                                                 Host-based markers for the early stage of infection by the grapevine trunk pathogen Neofusicoccum parvum
#> 122                                                                                                          RNASeq analysis of Vitis vinifera (cv. Sauvignon blanc) grape berries exposed to elevated light
#> 123                                                                                                           Dissecting the Variations of Transcriptome in Grape Berries Grown under Double Cropping System
#> 124                                                                                                                           Grape skin transcriptome to reveal genes related with resveratrol accumulation
#> 125                                                                             Shoot transcriptomic response is more sensitive to a 14 day water deficit than root transcriptomic response in Vitis riparia
#> 126                                                                           Dissecting the Variations of Ripening Progression and Flavonoid Metabolism in Grape Berries Grown under Double Cropping System
#> 127                                                               Whole-transcriptome microarray analysis of grapevine-Botrytis cinerea interaction during latent infection of berries (\xe2\xacSnoble rot⬝)
#> 128                                                                                                                    Effect of different irrigation supply on gene expression of two table grape cultivars
#> 129                                                                                                                            Analysis of gene expression in grape berries due to diurnal effects (DIU_05).
#> 130                                                                                                                                                        Rootstock and soil effects on grape transcriptome
#> 131                                                                                  Grape berry developmental series comparing irrigation treatments in a vineyard in Wingara, Victoria, Australia (WIN-05)
#> 132                                                                                            Root transcriptome comparison between grapevines Cabernet Sauvignon, 140 Ruggeri and K51-40 (+/- salt stress)
#> 133                                                                                          Effect of high CO2 levels and low temperature storage on table grapes quality harvested at two maturity stages.
#> 134                                                                                Vitis vinifera cultivar:Cabernet Sauvignon, Chardonnay, Merlot, Pinot Noir, and Semillon Transcriptome or Gene expression
#> 135                                                                                                                                                                   Vitis Transcriptome or Gene expression
#> 136                                                                                               Multi-omic characterization of the Vitis vinifera/Plasmopara viticola (Grapevine Downy Mildew) pathosystem
#> 137                                                                                                                                                     Bordeaux Cabernet Sauvignon berry skin transcriptome
#> 138                                                                                                                                                          Vitis vinifera Transcriptome or Gene expression
#> 139                                                                                                                  detection of messenger RNA exchange between scions and rootstocks in grafted grapevines
#> 140                                                                                                                                                 Transcriptional analysis of  the treatment of riboflavin
#> 141                                                                                                                                                  Vitis Pseudoreticulata Transcriptome or Gene expression
#> 142                                                                                                                                                          Vitis vinifera Transcriptome or Gene expression
#> 143                                                                                                                                     Vitis vinifera cultivar:Tempranillo Transcriptome or Gene expression
#> 144                                                                                                                                                                 Transcriptome of Vitis quinquangularis S
#> 145                                                                                                                       Molecular analysis of interaction between the grapevine Berry and Botrytis cinerea
#> 146                                                                                                           Vitis vinifera cultivar:Chasselas Blanc and Chasselas Cioutat Transcriptome or Gene expression
#> 147                                                                                                                                                                                           Vitis grafting
#> 148                                                                                                                                                                               Grape bud dormancy RNA-seq
#> 149                                                                                                                                                                            Vitis vinifera cv Tempranillo
#> 150                                                                                                  Comparative transcriptome analyses of grapevine inflorescences of three different developmental stages.
#> 151                                                                                                                                                                             GRBaV transcriptome analysis
#> 152                                                                                                                                                      Vitis vinifera cultivar:Variable Raw sequence reads
#> 153                                                                                                                   Transcriptome analysis of light effect on anthocyanin accumulation in grape berry skin
#> 154                                                                                                                                                Transcriptome of over-expression  grape lines for VpSTS29
#> 155                                                    Differences in berry primary and secondary metabolisms identified by transcriptomic and metabolic profiling of two table grape color somatic variants
#> 156                                                                                                                                                        Vitis vinifera cultivar:Merlot Raw sequence reads
#> 157                                                                                                                                                                 viromic study of P70 accession grapevine
#> 158                                                                                                                                                                         Vitis vinifera leaf freeze shock
#> 159                                                                                                                                                   Vitis vinifera cv Thompson Seedless Raw sequence reads
#> 160                                                                                              Comparison of two different host plant genera responding to grapevine leafroll-associated virus 3 infection
#> 161                                                                                                                                                                 Vitis vinifera Grape Berry Transcriptome
#> 162                                                                                                                                  Vitis vinifera strain:superior, 3003, Muscat Hamburg Raw sequence reads
#> 163                                                                                                                                                          Vitis vinifera Transcriptome or Gene expression
#> 164                                                                                                                   Transcriptomic comparison of four vitis species in response to water and salt stresses
#> 165                                                                                   Identification of the molecular regulators of softening and early ripening events in grape berries (Vitis vinifera L.)
#> 166                                                                                                                                                          Vitis vinifera Transcriptome or Gene expression
#> 167                                                                                                                                        Vitis vinifera cultivar:Loureira Transcriptome or Gene expression
#> 168                                                                                                                                                          Vitis vinifera Transcriptome or Gene expression
#> 169                                                                                                                                                     vitis vinifera shine muscat grape Raw sequence reads
#> 170                                                                                                                                                           Vitis davidii Transcriptome or Gene expression
#> 171                                                                                                                               Vitis vinifera cultivar:thompson seedless Transcriptome or Gene expression
#> 172                                                                                                                                       Vitis vinifera cultivar:Brachetto Transcriptome or Gene expression
#> 173                                                                                                                                                     Vitis amurensis isolate:Tonghua-3 Raw sequence reads
#> 174                                                                                                              Expression data from micropropagated Vitis vinifera when transferred to ex vitro conditions
#> 175                                                                                                                                                          Grapevine response to Planococcus ficus feeding
#> 176                                                                                                                                             Transcriptome profiles of three Muscat table grape cultivars
#> 177                                                                                                                                    Vitis vinifera cultivar:Summer Black Transcriptome or Gene expression
#> 178                                                                                        Gene expression patterns associated with grapevine resistance to downy mildew mediated by the Rpv1 and Rpv2 genes
#> 179                                                                                                                                                                 Vitis vinifera strain:Cabernet sauvignon
#> 180                                                                                                                                    Expression data in individual grape berries during ripening initation
#> 181                                                                                                                                                             Photoperiod regulation of grape bud dormancy
#> 182                                                                                      Transcriptomic analysis during heat stress and after the following recovery in grapevine (Vitis vinifera L.) leaves
#> 183                                                                                                            Gene expression in leaves of grapevine plants under abiotic stress conditions: heat and water
#> 184                          Comparative Physiological and Transcriptome Profiling of Grapevine Rootstocks Identified Their Physiological Adaptations and Gene Regulation  Networks behind Drought Tolerance
#> 185                                                                                                                                                   Grape flesh and skin ripening transcriptomic profiling
#> 186                                                                                     The grapevine expression atlas reveals a deep transcriptome shift driving the entire plant into a maturation program
#> 187                                                             Increasing the source/sink ratio in Vitis vinifera (cv Sangiovese) induces extensive transcriptome reprogramming and modifies berry ripening
#> 188                                                                                                     RNAs expression profiling in a resistant and a susceptible grapevine cultivar to Plasmopara viticola
#> 189                                                                            Transcript and metabolite analysis in Trincadeira cultivar reveals novel information regarding the dynamics of grape ripening
#> 190                                                                                                                                                                                 grape Raw sequence reads
#> 191                                                                                                                                           Time course of grapevine tendril and inflorescence development
#> 192                                                                                                                                      Transcriptional analysis of late ripening stages of grapevine berry
#> 193                                                                                         Transcriptome analysis along berry development of 'Syrah' x 'Pinot Noir' offspring extreme for flavonoid content
#> 194                                                  Five Omic Technologies Are Concordant in Differentiating the Biochemical Characteristics of the Berries of Five Grapevine (Vitis vinifera L.) Cultivars
#> 195                                                                                                         Disclosing the Molecular Basis of the Postharvest Life of Berry in Different Grapevine Genotypes
#> 196                Genome-wide expression profiling of Neofusicoccum parvum virulence factors in response to different growth substrates in vitro and during an interaction time course with grapevine stems
#> 197                                                                                                                Timing and order of the molecular events marking the onset of berry ripening in grapevine
#> 198                                                                                    Transcriptomic analysis of a white-to-red skin color reversion in grapevine (Vitis vinifera L. cv. ´Moscatel Galego´)
#> 199                                                                                                                                   Transcriptome analysis for shoot of Japanese wine grape cultivar Koshu
#> 200                                                                                                                                   Vitis vinifera cultivar:Prieto Picudo Transcriptome or Gene expression
#> 201                                                        Volatile profiles and transcriptomic variation in Cabernet Sauvignon grapes in response to various sunlight exposure strategies in a warm climate
#> 202                                                                 Using the combined analysis of transcripts and metabolites to propose key genes for differential terpene accumulation across two regions
#> 203                                                                                                           Grapevine cell cultures transcriptomic response to methyl jasmonate and cyclodextrin elicitors
#> 204                                                                                                                       Vitis vinifera isolate:latent bud | cultivar:Cabernet Sauvignon Raw sequence reads
#> 205                                                                                                                     The effects of heat shock treatment on gene expression of grapevine berries (HSP_06)
#> 206                                                                                                                                            Vitis vinifera strain:Tannat Transcriptome or Gene expression
#> 207                                                                                                                                                          Vitis vinifera Transcriptome or Gene expression
#> 208                                                                                                                                           Vitis vinifera strain:Corvina Transcriptome or Gene expression
#> 209                                                                                              Next Generation Sequencing of grapevine wood pathogen infection in the presence of a root biocontrol agent.
#> 210                                                                                                                                         Vitis vinifera root transcriptome analysis under flooding stress
#> 211                                                                                                                Vitis vinifera cultivar:Corinto Bianco and Pedro Ximenez Transcriptome or Gene expression
#> 212                                                                                                                  Vitis labrusca x Vitis vinifera strain:Labruscan grape Transcriptome or Gene expression
#> 213                                                                                                                                           Vitis vinifera transcriptome in response to Xylella fastidiosa
#> 214                           Grapevine comparative early transcriptome profiling reveals that Flavescence dor\xe9e phytoplasma represses plant responses induced by vector feeding in susceptible varieties
#> 215                                                                                             Transcriptional profiling of Vitis vinifera during grapevine-downy mildew interaction using cDNA microarrays

get_experiments(sampleName="GSM671721.ch1")

#>   experimentAccessId                                  experimentName
#> 1           GSE31674 Pinot Noir berry transcriptome during ripening.