get annotations for n samples from the selected compendium

get annotations for n samples from the selected compendium

get_sample_annotation(compendium = "vespucci", n = 10)

Arguments

compendium
A string - the selected compendium

n
an integer: number of sample to retrieve (default 10)

Value

A data.frame with three columns: sampleId sampleName annotation

Examples

get_sample_annotation(n=25)

#>               sampleId      sampleName
#> 1  U2FtcGxlVHlwZTozMjAw GSM2627902.ch1
#> 2  U2FtcGxlVHlwZToyMw==  GSM147690.ch1
#> 3  U2FtcGxlVHlwZToyMw==  GSM147690.ch1
#> 4  U2FtcGxlVHlwZToyMw==  GSM147690.ch1
#> 5  U2FtcGxlVHlwZToyMw==  GSM147690.ch1
#> 6  U2FtcGxlVHlwZToyMw==  GSM147690.ch1
#> 7  U2FtcGxlVHlwZTo1OQ==  GSM147726.ch1
#> 8  U2FtcGxlVHlwZTo1OQ==  GSM147726.ch1
#> 9  U2FtcGxlVHlwZTo1OQ==  GSM147726.ch1
#> 10 U2FtcGxlVHlwZTo1OQ==  GSM147726.ch1
#> 11 U2FtcGxlVHlwZTo1OQ==  GSM147726.ch1
#> 12 U2FtcGxlVHlwZTo3Nw==  GSM786240.ch1
#> 13 U2FtcGxlVHlwZTo3Nw==  GSM786240.ch1
#> 14 U2FtcGxlVHlwZTo3Nw==  GSM786240.ch1
#> 15 U2FtcGxlVHlwZTo3Nw==  GSM786240.ch1
#> 16 U2FtcGxlVHlwZTo3Nw==  GSM786240.ch1
#> 17 U2FtcGxlVHlwZTo3Nw==  GSM786240.ch1
#> 18 U2FtcGxlVHlwZTo3Nw==  GSM786240.ch1
#> 19 U2FtcGxlVHlwZTo3Nw==  GSM786240.ch1
#> 20 U2FtcGxlVHlwZTo3Nw==  GSM786240.ch1
#> 21 U2FtcGxlVHlwZTo3Nw==  GSM786240.ch1
#> 22 U2FtcGxlVHlwZTo4Ng==  GSM786249.ch1
#> 23 U2FtcGxlVHlwZTo4Ng==  GSM786249.ch1
#> 24 U2FtcGxlVHlwZTo4Ng==  GSM786249.ch1
#> 25 U2FtcGxlVHlwZTo4Ng==  GSM786249.ch1
#>                                                                                                                                                                                                                                                                                                                                                         annotation
#> 1                                                                                                                                       {"@id": "_:Na40b86b0d5674343a220a9232de33222", "@type": ["https://www.awri.com.au/wp-content/uploads/grapegrowth.pdf#35"], "http://purl.obolibrary.org/obo/NCIT_C25330": [{"@id": "_:Nc274a0c72ad449b7bed4d1c26fe82c0f"}]}
#> 2                                                                                                                                                                                                                                                                                            {"@id": "23", "@type": ["http://purl.obolibrary.org/obo/PO_0025034"]}
#> 3                                                                                                                                                             {"@id": "23", "http://www.ebi.ac.uk/efo/EFO_0005136": [{"@value": "Cabernet-Sauvignon"}], "http://purl.obolibrary.org/obo/NCIT_C16631": [{"@id": "http://purl.obolibrary.org/obo/NCBITaxon_29760"}]}
#> 4                                                                                                                                                                                                                                               {"@id": "23", "http://www.bioassayontology.org/bao#BAO_0002924": [{"@id": "_:N2d9927bae3de4e06869ce4e73609766e"}]}
#> 5                                                                                                                                                                                                                                                     {"@id": "23", "http://purl.obolibrary.org/obo/TO_0000541": [{"@id": "_:N5d5f93f04bbc4a42ba69485eeccf03f6"}]}
#> 6                                                                                                                                                                                {"@id": "_:N5d5f93f04bbc4a42ba69485eeccf03f6", "@type": ["http://purl.obolibrary.org/obo/NCIT_C104788"], "http://www.w3.org/1999/02/22-rdf-syntax-ns#value": [{"@value": "3+4"}]}
#> 7                                                                                                                                                                                                                                                                                            {"@id": "59", "@type": ["http://purl.obolibrary.org/obo/PO_0025034"]}
#> 8                                                                                                                                                                          {"@id": "59", "http://www.ebi.ac.uk/efo/EFO_0005136": [{"@value": "Norton"}], "http://purl.obolibrary.org/obo/NCIT_C16631": [{"@id": "http://purl.obolibrary.org/obo/NCBITaxon_3605"}]}
#> 9                                                                                                                                                                                                                                               {"@id": "59", "http://www.bioassayontology.org/bao#BAO_0002924": [{"@id": "_:Nce5b27efb7b54c869201c68520da6f35"}]}
#> 10                                                                                                                                                                                                                                                    {"@id": "59", "http://purl.obolibrary.org/obo/TO_0000541": [{"@id": "_:N21e187a168b74adca9733457de8755c0"}]}
#> 11                                                                                                                                                                               {"@id": "_:N21e187a168b74adca9733457de8755c0", "@type": ["http://purl.obolibrary.org/obo/NCIT_C104788"], "http://www.w3.org/1999/02/22-rdf-syntax-ns#value": [{"@value": "3+4"}]}
#> 12                                                                                                                                                                                                                          {"@id": "77", "http://purl.obolibrary.org/obo/PO_0007033": [{"@id": "https://www.awri.com.au/wp-content/uploads/grapegrowth.pdf#33"}]}
#> 13                                                                                                                                                                                                                                                   {"@id": "77", "http://purl.obolibrary.org/obo/NCIT_C99800": [{"@id": "_:N9035ea19a82a4c759b09771a246cb3c3"}]}
#> 14                                                                      {"@id": "_:N9035ea19a82a4c759b09771a246cb3c3", "http://purl.obolibrary.org/obo/UO_0000000": [{"@id": "http://purl.obolibrary.org/obo/UO_0000036"}], "http://www.w3.org/1999/02/22-rdf-syntax-ns#value": [{"@type": "http://www.w3.org/2001/XMLSchema#positiveInteger", "@value": "2003"}]}
#> 15                                                                                                                                                                    {"@id": "77", "http://www.ebi.ac.uk/efo/EFO_0005136": [{"@value": "Pinot Noir"}], "http://purl.obolibrary.org/obo/NCIT_C16631": [{"@id": "http://purl.obolibrary.org/obo/NCBITaxon_29760"}]}
#> 16                                                                                                                                                                                                                                                                                           {"@id": "77", "@type": ["http://purl.obolibrary.org/obo/PO_0009001"]}
#> 17                                                                                                                                                                                                                                              {"@id": "77", "http://www.bioassayontology.org/bao#BAO_0002924": [{"@id": "_:N6f3c098ba2294b84883ec5f47323be9c"}]}
#> 18                                                                                                                                                      {"@id": "_:N6f3c098ba2294b84883ec5f47323be9c", "@type": ["http://purl.obolibrary.org/obo/ENVO_00000116"], "http://purl.obolibrary.org/obo/GAZ_00000448": [{"@id": "_:N1fc99c1d085841de8b0eaf6fc51cdf4c"}]}
#> 19 {"@id": "_:N1fc99c1d085841de8b0eaf6fc51cdf4c", "http://purl.obolibrary.org/obo/NCIT_C68642": [{"@id": "_:Naf56ba2ce56046d6aed156ff406fedbf"}], "http://purl.obolibrary.org/obo/NCIT_C68643": [{"@id": "_:Nab6e43fd86474951aa2b166e186c668d"}], "http://www.w3.org/1999/02/22-rdf-syntax-ns#Description": [{"@value": "Ai molini San Michele all'Adige Italy"}]}
#> 20                                                                           {"@id": "_:Naf56ba2ce56046d6aed156ff406fedbf", "http://purl.obolibrary.org/obo/UO_0000000": [{"@id": "http://purl.obolibrary.org/obo/UO_0000185"}], "http://www.w3.org/1999/02/22-rdf-syntax-ns#value": [{"@type": "http://www.w3.org/2001/XMLSchema#float", "@value": "46.192921"}]}
#> 21                                                                           {"@id": "_:Nab6e43fd86474951aa2b166e186c668d", "http://purl.obolibrary.org/obo/UO_0000000": [{"@id": "http://purl.obolibrary.org/obo/UO_0000185"}], "http://www.w3.org/1999/02/22-rdf-syntax-ns#value": [{"@type": "http://www.w3.org/2001/XMLSchema#float", "@value": "11.134089"}]}
#> 22                                                                                                                                                                                                                          {"@id": "86", "http://purl.obolibrary.org/obo/PO_0007033": [{"@id": "https://www.awri.com.au/wp-content/uploads/grapegrowth.pdf#33"}]}
#> 23                                                                                                                                                                                                                                                   {"@id": "86", "http://purl.obolibrary.org/obo/NCIT_C99800": [{"@id": "_:N3c5d346bcbd6423680ba678330f9fb90"}]}
#> 24                                                                      {"@id": "_:N3c5d346bcbd6423680ba678330f9fb90", "http://purl.obolibrary.org/obo/UO_0000000": [{"@id": "http://purl.obolibrary.org/obo/UO_0000036"}], "http://www.w3.org/1999/02/22-rdf-syntax-ns#value": [{"@type": "http://www.w3.org/2001/XMLSchema#positiveInteger", "@value": "2005"}]}
#> 25                                                                                                                                                                    {"@id": "86", "http://www.ebi.ac.uk/efo/EFO_0005136": [{"@value": "Pinot Noir"}], "http://purl.obolibrary.org/obo/NCIT_C16631": [{"@id": "http://purl.obolibrary.org/obo/NCBITaxon_29760"}]}