get annotations for n samples from the selected compendium

get annotations for n samples from the selected compendium

get_sample_annotation(compendium = "vitis_vinifera", n = 10)

Arguments

compendium
a character - the selected compendium

n
an integer: number of sample to retrieve (default 10)

Value

a data.frame

Examples

get_sample_annotation(n=25)

#>       sampleName                 ontology.originalId               name
#> 1  GSM287866.ch1                                2006 Condition ontology
#> 2  GSM287866.ch1         GENOTYPE.CABERNET_SAUVIGNON Condition ontology
#> 3  GSM287866.ch1                  EICHHORN-LORENZ.33 Condition ontology
#> 4  GSM287866.ch1       PO:0008002_fruit_proximal_end Condition ontology
#> 5  GSM287866.ch1          VINEYARD.OSOYOOS_BC_CANADA Condition ontology
#> 6  GSM299462.ch1                       LEAF_POSITION Condition ontology
#> 7  GSM299462.ch1                           PLANT_AGE Condition ontology
#> 8  GSM299462.ch1                          GREENHOUSE Condition ontology
#> 9  GSM299462.ch1    GENOTYPE.MUSCADINIA_ROTUNDIFOLIA Condition ontology
#> 10 GSM299462.ch1                     PO:0025034_leaf Condition ontology
#> 11 GSM322364.ch1                                2006 Condition ontology
#> 12 GSM322364.ch1                   GENOTYPE.MACCABEU Condition ontology
#> 13 GSM322364.ch1                       LEAF_POSITION Condition ontology
#> 14 GSM322364.ch1                PO:0020138_leaf_vein Condition ontology
#> 15 GSM322364.ch1 VINEYARD.COVIOLO-REGGIOEMILIA_ITALY Condition ontology
#> 16 GSM322364.ch1                     VINEYARD.UNKOWN Condition ontology
#> 17 GSM322369.ch1                                2006 Condition ontology
#> 18 GSM322369.ch1                    GENOTYPE.MANZONI Condition ontology
#> 19 GSM322369.ch1                       LEAF_POSITION Condition ontology
#> 20 GSM322369.ch1                PO:0020138_leaf_vein Condition ontology
#> 21 GSM322369.ch1 VINEYARD.COVIOLO-REGGIOEMILIA_ITALY Condition ontology
#> 22 GSM322369.ch1                     VINEYARD.UNKOWN Condition ontology
#> 23 GSM404719.ch2                                TIME Condition ontology
#> 24 GSM404719.ch2                                2001 Condition ontology
#> 25 GSM404719.ch2                   GENOTYPE.PERLETTE Condition ontology